requirements.txt
pypdf
pandas

# Vector store & embeddings
chromadb


# Gemini API
google-generativeai==0.7.2

# Utilities
python-dotenv==1.0.1